refactor(ai-worker): isolate GeraLanding backend clients by stage
